requirements.txt
nltk
praw
psycopg2-binary
python-dotenv
